+++ src/main/java/amp/awec/command/argtypes/ArgumentTypeBlockPattern.java
@@ -72,7 +72,7 @@ public class ArgumentTypeBlockPattern implements ArgumentType<BlockPattern> {
 
 	@Override
 	public <S> CompletableFuture<Suggestions> listSuggestions(CommandContext<S> context, SuggestionsBuilder builder) {
-		String remaining = builder.getRemainingLowerCase();
+		String remaining = builder.getRemaining();
 		int lastOperatorIndex = RegexUtil.lastIndexOfSet(remaining, "[,%]");
 		String suggestionPrefix = "";
 		if (lastOperatorIndex != -1) {